Refactor Habit model to remove unnecessary export and improve code structure
This commit removes the unnecessary export statement for the Habit model in the Habit.js file. It also improves the code structure by moving the export statement to the end of the file for consistency.

Refactor userRoutes.js to handle password validation correctly

This commit fixes the password validation logic in the userRoutes.js file. It replaces the commented out code with the proper bcrypt password comparison logic to ensure that the credentials are validated correctly.

Update schema definitions for createUserParams.json and updateUserBodySchema.json

This commit updates the schema definitions for the createUserParams.json and updateUserBodySchema.json files. It removes the 'id' field from the required properties list in createUserParams.json and removes the 'id' property from the updateUserBodySchema.json file.

Add dailyHabitRoutes.js to handle daily habit CRUD operations

This commit adds the dailyHabitRoutes.js file, which handles the CRUD operations for daily habits. It provides routes for creating a new daily habit, getting all daily habits for a user, updating a daily habit by ID, and deleting a daily habit by ID.

Update server.js to register dailyHabitRoutes

This commit updates the server.js file to register the dailyHabitRoutes. It adds the necessary import statement for dailyHabitRoutes.js and includes it in the fastify register function. (post daily habit doesnt work)